backend/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv export --format requirements-txt --output-file requirements.txt --no-dev --no-editable --frozen --no-hashes
aioboto3==15.2.0
    # via pdf-extraction-tool
aiobotocore==2.24.2
    # via aioboto3
aiofiles==25.1.0
    # via aioboto3
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.13.1
    # via aiobotocore
aioitertools==0.12.0
    # via aiobotocore
aiosignal==1.4.0
    # via aiohttp
aiosqlite==0.21.0
    # via pdf-extraction-tool
amqp==5.3.1
    # via
    #   kombu
    #   pdf-extraction-tool
annotated-types==0.7.0
    # via
    #   pdf-extraction-tool
    #   pydantic
anyio==4.11.0
    # via
    #   httpx
    #   openai
    #   pdf-extraction-tool
    #   starlette
asyncpg==0.30.0
    # via pdf-extraction-tool
attrs==25.4.0
    # via aiohttp
beautifulsoup4==4.14.2
    # via
    #   markdownify
    #   markitdown
billiard==4.2.2
    # via celery
boto3==1.40.18
    # via aiobotocore
botocore==1.40.18
    # via
    #   aiobotocore
    #   boto3
    #   s3transfer
camelot-py==1.0.9
    # via pdf-extraction-tool
celery==5.5.3
    # via pdf-extraction-tool
certifi==2025.10.5
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==2.0.0 ; platform_python_implementation != 'PyPy'
    # via cryptography
chardet==5.2.0
    # via camelot-py
charset-normalizer==3.4.4
    # via
    #   markitdown
    #   pdfminer-six
    #   requests
click==8.3.0
    # via
    #   camelot-py
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
    #   magika
    #   uvicorn
click-didyoumean==0.3.1
    # via celery
click-plugins==1.1.1.2
    # via celery
click-repl==0.3.0
    # via celery
colorama==0.4.6 ; sys_platform == 'win32'
    # via
    #   click
    #   loguru
    #   tqdm
coloredlogs==15.0.1
    # via onnxruntime
cryptography==46.0.3
    # via pdfminer-six
defusedxml==0.7.1
    # via markitdown
distro==1.9.0
    # via openai
et-xmlfile==2.0.0
    # via openpyxl
fastapi==0.119.1
    # via pdf-extraction-tool
flatbuffers==25.9.23
    # via onnxruntime
frozenlist==1.8.0
    # via
    #   aiohttp
    #   aiosignal
greenlet==3.2.4
    # via
    #   pdf-extraction-tool
    #   sqlalchemy
gunicorn==23.0.0
    # via pdf-extraction-tool
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via openai
humanfriendly==10.0
    # via coloredlogs
idna==3.11
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
jiter==0.11.1
    # via openai
jmespath==1.0.1
    # via
    #   aiobotocore
    #   boto3
    #   botocore
kombu==5.5.4
    # via celery
loguru==0.7.3
    # via pdf-extraction-tool
magika==0.6.2
    # via markitdown
markdownify==1.2.0
    # via markitdown
markitdown==0.1.3
    # via pdf-extraction-tool
mpmath==1.3.0
    # via sympy
multidict==6.7.0
    # via
    #   aiobotocore
    #   aiohttp
    #   yarl
numpy==2.2.6
    # via
    #   camelot-py
    #   magika
    #   onnxruntime
    #   opencv-python
    #   opencv-python-headless
    #   pandas
    #   pdf-extraction-tool
onnxruntime==1.20.1
    # via
    #   magika
    #   markitdown
openai==2.6.0
    # via pdf-extraction-tool
opencv-python==4.12.0.88
    # via pdf-extraction-tool
opencv-python-headless==4.12.0.88
    # via camelot-py
openpyxl==3.1.5
    # via camelot-py
packaging==25.0
    # via
    #   gunicorn
    #   kombu
    #   onnxruntime
    #   pdf-extraction-tool
    #   pytesseract
pandas==2.3.3
    # via camelot-py
passlib==1.7.4
    # via pdf-extraction-tool
pdfminer-six==20250506
    # via
    #   camelot-py
    #   pdf-extraction-tool
    #   pdfplumber
pdfplumber==0.11.7
    # via pdf-extraction-tool
pillow==11.3.0
    # via
    #   camelot-py
    #   pdf-extraction-tool
    #   pdfplumber
    #   pytesseract
prompt-toolkit==3.0.52
    # via click-repl
propcache==0.4.1
    # via
    #   aiohttp
    #   yarl
protobuf==6.33.0
    # via onnxruntime
psycopg2==2.9.11
    # via pdf-extraction-tool
psycopg2-binary==2.9.11
    # via pdf-extraction-tool
pycparser==2.23 ; implementation_name != 'PyPy' and platform_python_implementation != 'PyPy'
    # via cffi
pydantic==2.11.9
    # via
    #   fastapi
    #   openai
    #   pdf-extraction-tool
pydantic-core==2.33.2
    # via pydantic
pyjwt==2.9.0
    # via pdf-extraction-tool
pymupdf==1.26.4
    # via pdf-extraction-tool
pypdf==5.9.0
    # via camelot-py
pypdf2==3.0.1
    # via pdf-extraction-tool
pypdfium2==4.30.0
    # via
    #   camelot-py
    #   pdf-extraction-tool
    #   pdfplumber
pyreadline3==3.5.4 ; sys_platform == 'win32'
    # via humanfriendly
pytesseract==0.3.13
    # via pdf-extraction-tool
python-dateutil==2.9.0.post0
    # via
    #   aiobotocore
    #   botocore
    #   celery
    #   pandas
python-dotenv==1.1.1
    # via
    #   magika
    #   pdf-extraction-tool
python-multipart==0.0.20
    # via pdf-extraction-tool
pytz==2025.2
    # via pandas
redis==5.2.1
    # via pdf-extraction-tool
requests==2.32.5
    # via
    #   markitdown
    #   pdf-extraction-tool
s3transfer==0.13.1
    # via boto3
six==1.17.0
    # via
    #   markdownify
    #   python-dateutil
sniffio==1.3.1
    # via
    #   anyio
    #   openai
soupsieve==2.8
    # via beautifulsoup4
sqlalchemy==2.0.43
    # via pdf-extraction-tool
starlette==0.48.0
    # via fastapi
sympy==1.14.0
    # via onnxruntime
tabulate==0.9.0
    # via camelot-py
tqdm==4.67.1
    # via openai
typing-extensions==4.15.0
    # via
    #   aiosignal
    #   aiosqlite
    #   anyio
    #   beautifulsoup4
    #   fastapi
    #   openai
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   starlette
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
tzdata==2025.2
    # via
    #   kombu
    #   pandas
urllib3==2.5.0
    # via
    #   botocore
    #   requests
uvicorn==0.37.0
    # via pdf-extraction-tool
vine==5.1.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.14
    # via prompt-toolkit
win32-setctime==1.2.0 ; sys_platform == 'win32'
    # via loguru
wrapt==1.17.3
    # via aiobotocore
yarl==1.22.0
    # via aiohttp
